# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Book Store API",
...
    "contact": {
      "name": "Book Store Support"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:60385/"
    }
  ],
  "paths": {
    "/api/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get all books",
        "operationId": "GetBooks",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListDtoOfBookDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get book by ID",
        "operationId": "GetBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/admin/books": {
      "post": {
        "tags": [
          "Admin - Books"
        ],
        "summary": "Create a new book",
        "operationId": "CreateBook",
        "parameters": [
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Admin - Books"
        ],
        "summary": "Get all books",
        "operationId": "GetAllBooksAdmin",
        "parameters": [
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/books/{id}": {
      "put": {
        "tags": [
          "Admin - Books"
        ],
        "summary": "Update a book",
        "operationId": "UpdateBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Books"
        ],
        "summary": "Delete a book",
        "operationId": "SoftDeleteBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "api-version",
            "in": "header",
            "description": "API version. Example: `1.0`"
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred language for localized content (configurable via LocalizationOptions)"
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "description": "Correlation ID for tracking requests across the system"
          },
          {
            "name": "X-Causation-ID",
            "in": "header",
            "description": "Causation ID linking this request to its trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorDto": {
        "required": [
          "id",
          "name",
          "biography"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "biography": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "BookDto": {
        "required": [
          "id",
          "title",
          "isbn",
          "language",
          "languageName",
          "description",
          "publicationDate",
          "isPreRelease",
          "publisher",
          "authors",
          "categories"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "isbn": {
            "type": [
              "null",
              "string"
            ]
          },
          "language": {
            "type": "string"
          },
          "languageName": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "publicationDate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PartialDate"
              }
            ]
          },
          "isPreRelease": {
            "type": "boolean"
          },
          "publisher": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PublisherDto"
              }
            ]
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDto"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            }
          }
        }
      },
      "BookTranslationDto": {
        "required": [
          "description"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "CategoryDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateBookRequest": {
        "required": [
          "title",
          "isbn",
          "language",
          "translations",
          "publicationDate",
          "publisherId",
          "authorIds",
          "categoryIds"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "isbn": {
            "type": [
              "null",
              "string"
            ]
          },
          "language": {
            "type": "string"
          },
          "translations": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/BookTranslationDto"
            }
          },
          "publicationDate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PartialDate"
              }
            ]
          },
          "publisherId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "authorIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "PagedListDtoOfBookDto": {
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalItemCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDto"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "totalItemCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "pageCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "PartialDate": { },
...
      "PublisherDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateBookRequest": {
        "required": [
          "title",
          "isbn",
          "language",
          "translations",
          "publicationDate",
          "publisherId",
          "authorIds",
          "categoryIds"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "isbn": {
            "type": [
              "null",
              "string"
            ]
          },
          "language": {
            "type": "string"
          },
          "translations": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/BookTranslationDto"
            }
          },
          "publicationDate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PartialDate"
              }
            ]
          },
          "publisherId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "authorIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
    }
  },
...
}