# src/ApiService/BookStore.ApiService/Models/CategoryDto.cs
namespace BookStore.ApiService.Models;

/// <summary>
/// DTO for category information (localized based on Accept-Language header)
/// </summary>
public record CategoryDto(
    Guid Id,
    string Name);